+++ chartify/view/icons.py
