+++ chartify/view/progress_widget.py
