+++ chartify/view/view_functions.py
